requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.0
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
frozenlist==1.4.0
greenlet==3.0.1
idna==3.5
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
yarl==1.9.3
